tools/requirements.txt
supabase
pandas
python-dotenv
openpyxl
# pdfplumber # Uncomment when needed
